Move TensorSchema to _buffer_utils and rename the latter to _tensor_schema
